documented test failure: axum-login incorrect table information

+++ examples/postgres/migrations/20240326193508_init.sql
@@ -1,7 +1,7 @@
 -- Create users table.
 create table if not exists users
 (
-    id integer primary key not null,
+    id bigint primary key not null,
     username text not null unique,
     password text not null
 );
